# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI 文字游戏引擎 — 游玩、创作、社交全栈 API",
        "title": "WorkshopEngine API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/play/games": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页查询已发布游戏（支持标签/类型/排序）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-discovery"
                ],
                "summary": "游戏列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "逗号分隔标签过滤",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "游戏类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序：new（默认）| hot",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/games/worldbook/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查看已发布游戏的世界书词条（需游戏开启 allow_player_worldbook_view）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-discovery"
                ],
                "summary": "玩家世界书",
                "parameters": [
                    {
                        "type": "string",
                        "description": "游戏 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/games/{slug}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按 slug 或 UUID 查询已发布游戏详情（含 comment_config）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-discovery"
                ],
                "summary": "游戏详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "游戏 slug 或 UUID",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/play/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "列出玩家会话（支持 game_id/user_id 过滤）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-discovery"
                ],
                "summary": "会话列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "按游戏过滤",
                        "name": "game_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按用户过滤",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为指定游戏创建新会话（自动注入 first_mes，递增 play_count）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "play-discovery"
                ],
                "summary": "创建会话",
                "parameters": [
                    {
                        "description": "创建请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "game_id": {
                                    "type": "string"
                                },
                                "user_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "session_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}